migrate to dotnet core 3.1
