+++ Lab9-Test_Bots_DirectLine/code/Finished/PictureBot/PictureBot.csproj.user
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <NameOfLastUsedPublishProfile>testbotalexi - Web Deploy</NameOfLastUsedPublishProfile>
+  </PropertyGroup>
+</Project>
