+++ Lab9-Test_Bots_DirectLine/code/Finished/PictureBotDL/PictureBotDL.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 16.0.30204.135
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PictureBotDL", "PictureBotDL.csproj", "{080491CB-209F-47CE-A246-35DDABE44D9E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{080491CB-209F-47CE-A246-35DDABE44D9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{080491CB-209F-47CE-A246-35DDABE44D9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{080491CB-209F-47CE-A246-35DDABE44D9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{080491CB-209F-47CE-A246-35DDABE44D9E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1DCFDF4B-9CA2-42EC-895D-67C5518444CC}
+	EndGlobalSection
+EndGlobal
