--- Lab9-Test_Bots_DirectLine/code/Starter/PictureBot.csproj.user
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <NameOfLastUsedPublishProfile>picturebotcjg1 - Web Deploy</NameOfLastUsedPublishProfile>
-  </PropertyGroup>
-</Project>
