--- Lab9-Test_Bots_DirectLine/code/Starter/PictureBot.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29102.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PictureBot", "PictureBot.csproj", "{3E368FFC-DE99-47F1-89DE-6689F7612209}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3E368FFC-DE99-47F1-89DE-6689F7612209}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3E368FFC-DE99-47F1-89DE-6689F7612209}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3E368FFC-DE99-47F1-89DE-6689F7612209}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3E368FFC-DE99-47F1-89DE-6689F7612209}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4172B3C6-1C89-40C4-95D6-35E925706100}
-	EndGlobalSection
-EndGlobal
