+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/DeploymentTemplates/new-rg-parameters.json
