+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/DeploymentTemplates/preexisting-rg-parameters.json
