+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/Middleware/IntentRecognizerMiddleware.cs
