+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/Models/ImageMapper.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace PictureBot.Models
 {
-    public class ImageMapper 
+    public class ImageMapper
     {
         public static SearchHit ToSearchHit(dynamic hit)
         {
