+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/PictureBot.csproj
@@ -7,7 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="Azure.AI.TextAnalytics" Version="5.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.9" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
     <PackageReference Include="Microsoft.Bot.Builder.AI.Luis" Version="4.10.3" />
     <PackageReference Include="Microsoft.Bot.Builder.Azure.Blobs" Version="4.10.3" />
     <PackageReference Include="Microsoft.Bot.Builder.Dialogs" Version="4.10.3" />
@@ -19,10 +19,5 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Properties\PublishProfiles\" />
-      <Folder Include="Properties\ServiceDependencies\" />
-    </ItemGroup>
 
 </Project>
