+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/PictureBot.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30503.244
+MinimumVisualStudioVersion = 16.0.30503.244
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PictureBot", "PictureBot.csproj", "{C987F603-9E4A-49FA-A33F-CBEEF1CC1051}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C987F603-9E4A-49FA-A33F-CBEEF1CC1051}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C987F603-9E4A-49FA-A33F-CBEEF1CC1051}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C987F603-9E4A-49FA-A33F-CBEEF1CC1051}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C987F603-9E4A-49FA-A33F-CBEEF1CC1051}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E7C49197-C8CF-43BE-A58F-83656202CDA4}
+	EndGlobalSection
+EndGlobal
