+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/PictureBotAccessors.cs
@@ -1,0 +1,52 @@
+﻿using System;
+using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
+
+namespace Microsoft.PictureBot
+{
+    /// <summary>
+    /// This class is created as a Singleton and passed into the IBot-derived constructor.
+    ///  - See <see cref="PictureBot"/> constructor for how that is injected.
+    ///  - See the Startup.cs file for more details on creating the Singleton that gets
+    ///    injected into the constructor.
+    /// </summary>
+    public class PictureBotAccessors
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PictureBotAccessors"/> class.
+        /// Contains the <see cref="ConversationState"/> and associated <see cref="IStatePropertyAccessor{T}"/>.
+        /// </summary>
+        /// <param name="conversationState">The state object that stores the counter.</param>
+        public PictureBotAccessors(ConversationState conversationState)
+        {
+            ConversationState = conversationState ?? throw new ArgumentNullException(nameof(conversationState));
+        }
+
+        /// <summary>
+        /// Gets the <see cref="IStatePropertyAccessor{T}"/> name used for the <see cref="CounterState"/> accessor.
+        /// </summary>
+        /// <remarks>Accessors require a unique name.</remarks>
+        /// <value>The accessor name for the counter accessor.</value>
+        public static string PictureStateName { get; } = $"{nameof(PictureBotAccessors)}.PictureState";
+
+        /// <summary>
+        /// Gets or sets the <see cref="IStatePropertyAccessor{T}"/> for CounterState.
+        /// </summary>
+        /// <value>
+        /// The accessor stores the turn count for the conversation.
+        /// </value>
+        public IStatePropertyAccessor<PictureState> PictureState { get; set; }
+
+        /// <summary>
+        /// Gets the <see cref="ConversationState"/> object for the conversation.
+        /// </summary>
+        /// <value>The <see cref="ConversationState"/> object.</value>
+        public ConversationState ConversationState { get; }
+
+        /// <summary> Gets the IStatePropertyAccessor{T} name used for the DialogState accessor. </summary>
+        public static string DialogStateName { get; } = $"{nameof(PictureBotAccessors)}.DialogState";
+
+        /// <summary> Gets or sets the IStatePropertyAccessor{T} for DialogState. </summary>
+        public IStatePropertyAccessor<DialogState> DialogStateAccessor { get; set; }
+    }
+}
