+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/Properties/launchSettings.json
