+++ Lab9-Test_Bots_DirectLine/code/Starter/PictureBot/Startup.cs
@@ -1,25 +1,29 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 //
-// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.5.0
+// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.10.3
 
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+using PictureBot.Bots;
 
 using System;
 using System.Linq;
 using System.Text.RegularExpressions;
-using Microsoft.Bot.Builder.AI.Luis;
-using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Integration;
+using Microsoft.Bot.Configuration;
 using Microsoft.Bot.Connector.Authentication;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Logging;
+
+using Microsoft.Bot.Builder.AI.Luis;
+using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.PictureBot;
 using Microsoft.Bot.Builder.Azure.Blobs;
 using Azure.AI.TextAnalytics;
@@ -46,16 +50,27 @@ namespace PictureBot
             // Create the Bot Framework Adapter with error handling enabled.
             services.AddSingleton<IBotFrameworkHttpAdapter, AdapterWithErrorHandler>();
 
-            services.AddSingleton<IStorage, BlobsStorage>(sp =>
+            // Create the bot as a transient. In this case the ASP Controller is expecting an IBot.
+            services.AddBot<PictureBot.Bots.PictureBot>(options =>
             {
+                var appId = Configuration.GetSection("MicrosoftAppId")?.Value;
+                var appSecret = Configuration.GetSection("MicrosoftAppPassword")?.Value;
+
+                options.CredentialProvider = new SimpleCredentialProvider(appId, appSecret);
+
+                // Creates a logger for the application to use.
+                ILogger logger = _loggerFactory.CreateLogger<PictureBot.Bots.PictureBot>();
+
+                // Catches any errors that occur during a conversation turn and logs them.
+                options.OnTurnError = async (context, exception) =>
+                {
+                    logger.LogError($"Exception caught : {exception}");
+                    await context.SendActivityAsync("Sorry, it looks like something went wrong.");
+                };
+
                 // The Memory Storage used here is for local bot debugging only. When the bot
                 // is restarted, everything stored in memory will be gone.
-                //IStorage dataStore = new MemoryStorage();
-
-                //var blobConnectionString = Configuration.GetSection("BlobStorageConnectionString")?.Value;
-                //var blobContainer = Configuration.GetSection("BlobStorageContainer")?.Value;
-                //IStorage dataStore = new BlobsStorage(blobConnectionString, blobContainer);
-                //services.AddSingleton<IStorage>(dataStore);
+                IStorage dataStore = new MemoryStorage();
 
                 // For production bots use the Azure Blob or
                 // Azure CosmosDB storage providers. For the Azure
@@ -75,56 +90,29 @@ namespace PictureBot
                 // var storageContainer = string.IsNullOrWhiteSpace(blobStorageConfig.Container) ? DefaultBotContainer : blobStorageConfig.Container;
                 // IStorage dataStore = new Microsoft.Bot.Builder.Azure.AzureBlobStorage(blobStorageConfig.ConnectionString, storageContainer);
 
+                // Create Conversation State object.
+                // The Conversation State object is where we persist anything at the conversation-scope.
+                var userState = new UserState(dataStore);
+                var conversationState = new ConversationState(dataStore);
 
-                var blobConnectionString = Configuration.GetSection("BlobStorageConnectionString")?.Value;
-                var blobContainer = Configuration.GetSection("BlobStorageContainer")?.Value;
-                BlobsStorage dataStore = new BlobsStorage(blobConnectionString, blobContainer);
-                return dataStore;
+                // Create the User state.
+                services.AddSingleton<UserState>(userState);
+
+                // Create the Conversation state.
+                services.AddSingleton<ConversationState>(conversationState);
+
+                var middleware = options.Middleware;
+                middleware.Add(new RegExpRecognizerMiddleware()
+                .AddIntent("search", new Regex("search picture(?:s)*(.*)|search pic(?:s)*(.*)", RegexOptions.IgnoreCase))
+                .AddIntent("share", new Regex("share picture(?:s)*(.*)|share pic(?:s)*(.*)", RegexOptions.IgnoreCase))
+                .AddIntent("order", new Regex("order picture(?:s)*(.*)|order print(?:s)*(.*)|order pic(?:s)*(.*)", RegexOptions.IgnoreCase))
+                .AddIntent("help", new Regex("help(.*)", RegexOptions.IgnoreCase)));
             });
-
-            // Create the User state.
-            services.AddSingleton<UserState>(sp => {
-                var dataStore = sp.GetRequiredService<IStorage>();
-                return new UserState(dataStore);
-            });
-
-            // Create the Conversation state.
-            services.AddSingleton<ConversationState>(sp =>
-            {
-                var dataStore = sp.GetRequiredService<IStorage>();
-                return new ConversationState(dataStore);
-            });
-
-            services.AddBot<PictureBot.Bots.PictureBot>(options =>
-            {
-                var appId = Configuration.GetSection("MicrosoftAppId")?.Value;
-                var appSecret = Configuration.GetSection("MicrosoftAppPassword")?.Value;
-
-                options.CredentialProvider = new SimpleCredentialProvider(appId, appSecret);
-
-                // Creates a logger for the application to use.
-                ILogger logger = _loggerFactory.CreateLogger<PictureBot.Bots.PictureBot>();
-
-                // Catches any errors that occur during a conversation turn and logs them.
-                options.OnTurnError = async (context, exception) =>
-                {
-                    logger.LogError($"Exception caught : {exception}");
-                    await context.SendActivityAsync("Sorry, it looks like something went wrong.");
-                };
-
-                options.Middleware.Add(new RegExpRecognizerMiddleware()
-                    .AddIntent("search", new Regex("search picture(?:s)*(.*)|search pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                    .AddIntent("share", new Regex("share picture(?:s)*(.*)|share pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                    .AddIntent("order", new Regex("order picture(?:s)*(.*)|order print(?:s)*(.*)|order pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                    .AddIntent("help", new Regex("help(.*)", RegexOptions.IgnoreCase)));
-
-            });
-
 
             // Create and register state accesssors.
             // Acessors created here are passed into the IBot-derived class on every turn.
             services.AddSingleton<PictureBotAccessors>(sp =>
-            { 
+            {
 
                 var options = sp.GetRequiredService<IOptions<BotFrameworkOptions>>().Value;
                 if (options == null)
@@ -149,13 +137,14 @@ namespace PictureBot
                 };
             });
 
+
             // Create and register a LUIS recognizer.
             services.AddSingleton(sp =>
             {
                 var luisApplication = new LuisApplication(
-                   Configuration.GetSection("luisAppId")?.Value,
-                   Configuration.GetSection("luisAppKey")?.Value,
-                   Configuration.GetSection("luisEndPoint")?.Value);
+                  Configuration.GetSection("luisAppId")?.Value,
+                  Configuration.GetSection("luisAppKey")?.Value,
+                  Configuration.GetSection("luisEndPoint")?.Value);
                 // Set the recognizer options depending on which endpoint version you want to use.
                 // More details can be found in https://docs.microsoft.com/en-gb/azure/cognitive-services/luis/luis-migration-api-v3
                 var recognizerOptions = new LuisRecognizerOptionsV3(luisApplication)
@@ -164,8 +153,31 @@ namespace PictureBot
                     {
                         IncludeAllIntents = true,
                     }
-                };            
+                };
                 return new LuisRecognizer(recognizerOptions);
+            });
+
+            // Create the User state.
+            services.AddSingleton<UserState>(sp =>
+            {
+                var dataStore = sp.GetRequiredService<IStorage>();
+                return new UserState(dataStore);
+            });
+
+            // Create the Conversation state.
+            services.AddSingleton<ConversationState>(sp =>
+            {
+                var dataStore = sp.GetRequiredService<IStorage>();
+                return new ConversationState(dataStore);
+            });
+
+            // Create the IStorage.
+            services.AddSingleton<IStorage, BlobsStorage>(sp =>
+            {
+                var blobConnectionString = Configuration.GetSection("BlobStorageConnectionString")?.Value;
+                var blobContainer = Configuration.GetSection("BlobStorageContainer")?.Value;
+                BlobsStorage dataStore = new BlobsStorage(blobConnectionString, blobContainer);
+                return dataStore;
             });
 
             services.AddSingleton<TextAnalyticsClient>(sp =>
@@ -178,6 +190,7 @@ namespace PictureBot
             });
         }
 
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, ILoggerFactory loggerFactory)
         {
             _loggerFactory = loggerFactory;
