--- Lab9-Test_Bots_DirectLine/code/Starter/PictureState.cs
@@ -1,24 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System.Collections.Generic;
-
-namespace Microsoft.PictureBot
-{
-    /// <summary>
-    /// Stores counter state for the conversation.
-    /// Stored in <see cref="Microsoft.Bot.Builder.ConversationState"/> and
-    /// backed by <see cref="Microsoft.Bot.Builder.MemoryStorage"/>.
-    /// </summary>
-    public class PictureState
-    {
-        /// <summary>
-        /// Gets or sets the number of turns in the conversation.
-        /// </summary>
-        /// <value>The number of turns in the conversation.</value>
-        public string Greeted { get; set; } = "not greeted";
-        public List<string> UtteranceList { get; private set; } = new List<string>();
-        public string Search { get; set; } = "";
-        public string Searching { get; set; } = "no";
-    }
-}
