--- Lab9-Test_Bots_DirectLine/code/Starter/Properties/PublishProfiles/picturebotcjg1 - Web Deploy.pubxml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-This file is used by the publish/package process of your Web project. You can customize the behavior of this process
-by editing this MSBuild file. In order to learn more about this please visit https://go.microsoft.com/fwlink/?LinkID=208121. 
--->
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <WebPublishMethod>MSDeploy</WebPublishMethod>
-    <ResourceId>/subscriptions/e433f371-e5e9-4238-abc2-7c38aa596a18/resourceGroups/cjg-learnai/providers/Microsoft.Web/sites/picturebotcjg1</ResourceId>
-    <ResourceGroup>cjg-learnai</ResourceGroup>
-    <PublishProvider>AzureWebSite</PublishProvider>
-    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish>http://picturebotcjg1.azurewebsites.net</SiteUrlToLaunchAfterPublish>
-    <LaunchSiteAfterPublish>True</LaunchSiteAfterPublish>
-    <ExcludeApp_Data>False</ExcludeApp_Data>
-    <ProjectGuid>3e368ffc-de99-47f1-89de-6689f7612209</ProjectGuid>
-    <MSDeployServiceURL>picturebotcjg1.scm.azurewebsites.net:443</MSDeployServiceURL>
-    <DeployIisAppPath>picturebotcjg1</DeployIisAppPath>
-    <RemoteSitePhysicalPath />
-    <SkipExtraFilesOnServer>True</SkipExtraFilesOnServer>
-    <MSDeployPublishMethod>WMSVC</MSDeployPublishMethod>
-    <EnableMSDeployBackup>True</EnableMSDeployBackup>
-    <UserName>$picturebotcjg1</UserName>
-    <_SavePWD>True</_SavePWD>
-    <_DestinationType>AzureWebSite</_DestinationType>
-    <InstallAspNetCoreSiteExtension>False</InstallAspNetCoreSiteExtension>
-  </PropertyGroup>
-</Project>
