--- Lab9-Test_Bots_DirectLine/code/Starter/Responses/MainResponses.cs
@@ -1,38 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.Bot.Builder;
-
-namespace PictureBot.Responses
-{
-    public class MainResponses
-    {
-        public static async Task ReplyWithGreeting(ITurnContext context)
-        {
-            // Add a greeting
-        }
-        public static async Task ReplyWithHelp(ITurnContext context)
-        {
-            await context.SendActivityAsync($"I can search for pictures, share pictures and order prints of pictures.");
-        }
-        public static async Task ReplyWithResumeTopic(ITurnContext context)
-        {
-            await context.SendActivityAsync($"What can I do for you?");
-        }
-        public static async Task ReplyWithConfused(ITurnContext context)
-        {
-            // Add a response for the user if Regex or LUIS doesn't know
-            // What the user is trying to communicate
-        }
-        public static async Task ReplyWithLuisScore(ITurnContext context, string key, double score)
-        {
-            await context.SendActivityAsync($"Intent: {key} ({score}).");
-        }
-        public static async Task ReplyWithShareConfirmation(ITurnContext context)
-        {
-            await context.SendActivityAsync($"Posting your picture(s) on twitter...");
-        }
-        public static async Task ReplyWithOrderConfirmation(ITurnContext context)
-        {
-            await context.SendActivityAsync($"Ordering standard prints of your picture(s)...");
-        }
-    }
-}
