--- Lab9-Test_Bots_DirectLine/code/Starter/Startup.cs
@@ -1,191 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-//
-// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.5.0
-
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Builder.Integration.AspNet.Core;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-using System;
-using System.Linq;
-using System.Text.RegularExpressions;
-using Microsoft.Bot.Builder.AI.Luis;
-using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.Integration;
-using Microsoft.Bot.Configuration;
-using Microsoft.Bot.Connector.Authentication;
-using Microsoft.Extensions.Options;
-using Microsoft.Extensions.Logging;
-using Microsoft.PictureBot;
-
-using Microsoft.Azure.CognitiveServices.Language.TextAnalytics;
-using Microsoft.Azure.CognitiveServices.Language.TextAnalytics.Models;
-using Microsoft.Azure.CognitiveServices.Language.LUIS.Runtime;
-
-namespace PictureBot
-{
-    public class Startup
-    {
-        private ILoggerFactory _loggerFactory;
-        private bool _isProduction = false;
-
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-
-            // Create the Bot Framework Adapter with error handling enabled.
-            services.AddSingleton<IBotFrameworkHttpAdapter, AdapterWithErrorHandler>();
-
-            services.AddBot<PictureBot.Bots.PictureBot>(options =>
-            {
-                var appId = Configuration.GetSection("MicrosoftAppId")?.Value;
-                var appSecret = Configuration.GetSection("MicrosoftAppPassword")?.Value;
-
-                options.CredentialProvider = new SimpleCredentialProvider(appId, appSecret);
-
-                // Creates a logger for the application to use.
-                ILogger logger = _loggerFactory.CreateLogger<PictureBot.Bots.PictureBot>();
-
-                // Catches any errors that occur during a conversation turn and logs them.
-                options.OnTurnError = async (context, exception) =>
-                {
-                    logger.LogError($"Exception caught : {exception}");
-                    await context.SendActivityAsync("Sorry, it looks like something went wrong.");
-                };
-
-                // The Memory Storage used here is for local bot debugging only. When the bot
-                // is restarted, everything stored in memory will be gone.
-                var blobConnectionString = Configuration.GetSection("BlobStorageConnectionString")?.Value;
-                var blobContainer = Configuration.GetSection("BlobStorageContainer")?.Value;
-                IStorage dataStore = new Microsoft.Bot.Builder.Azure.AzureBlobStorage(blobConnectionString, blobContainer);
-
-                // For production bots use the Azure Blob or
-                // Azure CosmosDB storage providers. For the Azure
-                // based storage providers, add the Microsoft.Bot.Builder.Azure
-                // Nuget package to your solution. That package is found at:
-                // https://www.nuget.org/packages/Microsoft.Bot.Builder.Azure/
-                // Uncomment the following lines to use Azure Blob Storage
-                // //Storage configuration name or ID from the .bot file.
-                // const string StorageConfigurationId = "<STORAGE-NAME-OR-ID-FROM-BOT-FILE>";
-                // var blobConfig = botConfig.FindServiceByNameOrId(StorageConfigurationId);
-                // if (!(blobConfig is BlobStorageService blobStorageConfig))
-                // {
-                //    throw new InvalidOperationException($"The .bot file does not contain an blob storage with name '{StorageConfigurationId}'.");
-                // }
-                // // Default container name.
-                // const string DefaultBotContainer = "botstate";
-                // var storageContainer = string.IsNullOrWhiteSpace(blobStorageConfig.Container) ? DefaultBotContainer : blobStorageConfig.Container;
-                // IStorage dataStore = new Microsoft.Bot.Builder.Azure.AzureBlobStorage(blobStorageConfig.ConnectionString, storageContainer);
-
-                // Create Conversation State object.
-                // The Conversation State object is where we persist anything at the conversation-scope.
-                var userState = new UserState(dataStore);
-                var conversationState = new ConversationState(dataStore);
-
-                // Create the User state.
-                services.AddSingleton<UserState>(userState);
-
-                // Create the Conversation state.
-                services.AddSingleton<ConversationState>(conversationState);
-
-                var middleware = options.Middleware;
-
-                // Add middleware below with "middleware.Add(...."
-                // Add Regex below
-                middleware.Add(new RegExpRecognizerMiddleware()
-                .AddIntent("search", new Regex("search picture(?:s)*(.*)|search pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                .AddIntent("share", new Regex("share picture(?:s)*(.*)|share pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                .AddIntent("order", new Regex("order picture(?:s)*(.*)|order print(?:s)*(.*)|order pic(?:s)*(.*)", RegexOptions.IgnoreCase))
-                .AddIntent("help", new Regex("help(.*)", RegexOptions.IgnoreCase)));
-            });
-
-            // Create and register state accesssors.
-            // Acessors created here are passed into the IBot-derived class on every turn.
-            services.AddSingleton<PictureBotAccessors>(sp =>
-            {
-                var options = sp.GetRequiredService<IOptions<BotFrameworkOptions>>().Value;
-                if (options == null)
-                {
-                    throw new InvalidOperationException("BotFrameworkOptions must be configured prior to setting up the state accessors");
-                }
-
-                var conversationState = services.BuildServiceProvider().GetService<ConversationState>();
-
-                if (conversationState == null)
-                {
-                    throw new InvalidOperationException("ConversationState must be defined and added before adding conversation-scoped state accessors.");
-                }
-
-                // Create the custom state accessor.
-                // State accessors enable other components to read and write individual properties of state.
-                var accessors = new PictureBotAccessors(conversationState)
-                {
-                    PictureState = conversationState.CreateProperty<PictureState>(PictureBotAccessors.PictureStateName),
-                    DialogStateAccessor = conversationState.CreateProperty<DialogState>("DialogState"),
-                };
-
-                return accessors;
-            });
-
-
-            // Create and register a LUIS recognizer.
-            services.AddSingleton(sp =>
-            {
-                var luisAppId = Configuration.GetSection("luisAppId")?.Value;
-                var luisAppKey = Configuration.GetSection("luisAppKey")?.Value;
-                var luisEndPoint = Configuration.GetSection("luisEndPoint")?.Value;
-
-                // Get LUIS information
-                var luisApp = new LuisApplication(luisAppId, luisAppKey, luisEndPoint);
-
-                // Specify LUIS options. These may vary for your bot.
-                var luisPredictionOptions = new LuisPredictionOptions
-                {
-                    IncludeAllIntents = true,
-                };
-
-                // Create the recognizer
-                var recognizer = new LuisRecognizer(luisApp, luisPredictionOptions, true, null);
-                return recognizer;
-            });
-
-            services.AddSingleton(sp =>
-            {
-                string cogsBaseUrl = Configuration.GetSection("cogsBaseUrl")?.Value;
-                string cogsKey = Configuration.GetSection("cogsKey")?.Value;
-
-                var credentials = new ApiKeyServiceClientCredentials(cogsKey);
-                TextAnalyticsClient client = new TextAnalyticsClient(credentials)
-                {
-                    Endpoint = cogsBaseUrl
-                };
-
-                return client;
-            });
-        }
-
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
-        {
-            _loggerFactory = loggerFactory;
-
-            app.UseDefaultFiles()
-                .UseStaticFiles()
-                .UseBotFramework();
-
-            app.UseMvc();
-        }
-    }
-}
