--- Lab9-Test_Bots_DirectLine/code/Starter/appsettings.json
@@ -1,11 +1,0 @@
-{
-  "MicrosoftAppId": "YourAppID",
-  "MicrosoftAppPassword": "YourAppIDKey",
-  "BlobStorageConnectionString": "DefaultEndpointsProtocol=https;AccountName=XXXXX;AccountKey=XXXXX;EndpointSuffix=core.windows.net",
-  "BlobStorageContainer": "chatlog",
-  "luisAppId": "YourLouisAppID",
-  "luisAppKey": "YourLouisAppKey",
-  "luisEndPoint": "https://westus.api.cognitive.microsoft.com",
-  "cogsBaseUrl": "https://westus2.api.cognitive.microsoft.com",
-  "cogsKey": "YourCServiceKey"
-}
